styling and bug fix

+++ src/components/Tweet.js
@@ -1,5 +1,7 @@
 import { dbService, storageService } from 'fbase';
 import React, { useState } from 'react'
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faTrash, faPencilAlt } from "@fortawesome/free-solid-svg-icons";
 
 function Tweet({ tweetObj, isOwner }) {
   const [editing, setEditing] = useState(false);
@@ -10,7 +12,9 @@ function Tweet({ tweetObj, isOwner }) {
     if (ok) {
       // delete
       await dbService.doc(`tweets/${tweetObj.id}`).delete();
-      await storageService.refFromURL(tweetObj.attachmentUrl).delete();
+      if(tweetObj.attachmentUrl !== ""){
+        await storageService.refFromURL(tweetObj.attachmentUrl).delete();
+      }
     }
   };
 
@@ -27,35 +31,41 @@ function Tweet({ tweetObj, isOwner }) {
     setNewTweet(value);
   };
   return (
-    <div>
+    <div className="nweet">
       {
         editing ? (
           <>
             {isOwner && 
             <>
-            <form onSubmit={onSubmit}>
+            <form onSubmit={onSubmit} className="container nweetEdit">
               <input 
                 type="text" 
                 placeholder="Edit your tweet" 
                 value={newTweet} 
                 required
                 onChange={onChange}
+                className="formInput"
               />
-              <input type="submit" value="Update Tweet" />
+              <input type="submit" value="Update Tweet" className="formBtn"/>
             </form>
-            <button onClick={toggleEditing}>Cancel</button>
+            <button onClick={toggleEditing} className="formBtn cancelBtn">Cancel</button>
             </>}
           </>
         ) : (
           <>
           <h4>{tweetObj.text}</h4>
-          {tweetObj.attachmentUrl && <img src={tweetObj.attachmentUrl} width="50px" height="50px" alt=""/>}
+          {tweetObj.attachmentUrl && <img src={tweetObj.attachmentUrl} alt=""/>}
           {isOwner && (
-            <>
-              <button onClick={onDeleteClick}>Delete Tweet</button>
-              <button onClick={toggleEditing}>Edit Tweet</button>
-            </>
+            <div className="nweet__actions">
+              <span onClick={onDeleteClick}>
+                <FontAwesomeIcon icon={faTrash} />
+              </span>
+              <span onClick={toggleEditing}>
+                <FontAwesomeIcon icon={faPencilAlt} />
+              </span>
+            </div>
           )}
+          <span style={{marginTop:" 10px", fontSize: "10px"}}>by {tweetObj.creatorDisplayName}</span>
           </>
         )
       }
